feat: automate creation of aws infrastructure
This simple configuration just builds a VM on AWS. It runs on a public subnet in its own VPC. The
machine type is just a t3.medium, which was just used to get something running. This is by no means
the optimal type of machine to run Odamex on. I'll find that out in due course.

There are some scripts that build odasrv and the odamex.wad. These in turn required building the
Deutex tool, which itself required a specific, newer version of autoconf, that wasn't available in
the apt repositories.

As of this commit, these scripts are just run manually, but shortly, we'll use Ansible to run them.

Using this configuration, I was able to run and connect to a server. There's no configuration for
the server, as of yet.

The README mentions the use of Ansible, but this hasn't actually been setup yet.